Integration Tests: group tests by tested entity
