+++ tests/cli.rs
@@ -1,564 +1,19 @@
 extern crate diesel;
 extern crate fluetl;
 
-use chrono::NaiveDate;
-use chrono::NaiveTime;
-use diesel::{sql_query, QueryableByName, RunQueryDsl};
+use diesel::{sql_query, RunQueryDsl};
 use std::env;
 use std::fs::File;
 use std::io::BufRead;
 use std::io::BufReader;
-use std::ops::Range;
 use std::path::Path;
-use std::process::Command;
 
 use fluetl::infrastructure::database::connection::{establish_connection_pool, DbConnection};
 
 use diesel_migrations::{embed_migrations, EmbeddedMigrations, MigrationHarness};
 pub const MIGRATIONS: EmbeddedMigrations = embed_migrations!("migrations");
 
-// ****************** //
-// test fluetl --env-file=.env.test import mapping-client
-// ******************
-
-#[test]
-fn import_mapping_client_once() {
-    import_mapping_client(0..1);
-}
-
-#[test]
-fn import_mapping_client_10_times() {
-    import_mapping_client(0..10);
-}
-
-fn import_mapping_client(repeat: Range<i32>) {
-    // Arrange
-    let mut connection = setup_database_connection();
-    reset_test_database(&mut connection);
-
-    insert_raw_sql("tests/fixtures/mapping_client_source.sql", &mut connection)
-        .expect("Failed to insert mapping_client_source.sql fixture file");
-
-    // Result
-    for _ in repeat {
-        let output = Command::new("target/debug/fluetl")
-            .args(["import", "mapping-client", "--env-file=.env.test"])
-            .output()
-            .expect("Failed to execute command");
-
-        if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            println!("Command executed successfully:\n{}", stdout);
-        } else {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            eprintln!("Command failed:\n{}", stderr);
-        }
-    }
-
-    //Assert
-    let query_results = sql_query("SELECT * From `mapping_client_contact`")
-        .load::<MappingClientPlaceholder>(&mut connection)
-        .expect("Failed to query mapping_client_contact table");
-
-    let query_result_sample_1 =
-        sql_query("SELECT * From `mapping_client_contact` WHERE idp_id_client = 1009681")
-            .load::<MappingClientPlaceholder>(&mut connection)
-            .expect("Failed to query mapping_client_contact table");
-
-    let query_result_sample_2 =
-        sql_query("SELECT * From `mapping_client_contact` WHERE idp_id_client = 1010265")
-            .load::<MappingClientPlaceholder>(&mut connection)
-            .expect("Failed to query mapping_client_contact table");
-
-    assert_eq!(query_results.len(), 4); // mapping_client_source.sql has 5 rows but one is ignored because of null id_customer
-    assert_eq!(
-        query_result_sample_1[0],
-        MappingClientPlaceholder {
-            id_customer: 1,
-            idp_id_client: 1009681
-        }
-    );
-    assert_eq!(
-        query_result_sample_2[0],
-        MappingClientPlaceholder {
-            id_customer: 4,
-            idp_id_client: 1010265
-        }
-    );
-}
-
-#[derive(QueryableByName, Debug, PartialEq)]
-struct MappingClientPlaceholder {
-    #[diesel(sql_type = diesel::sql_types::Integer)]
-    pub id_customer: i32,
-    #[diesel(sql_type = diesel::sql_types::Integer)]
-    pub idp_id_client: i32,
-}
-
-// ****************** //
-// test fluetl import order --env-file=.env.test
-// ****************** //
-#[test]
-fn import_order_once() {
-    import_order(0..1);
-}
-
-#[test]
-fn import_order_10_times() {
-    import_order(0..10);
-}
-
-fn import_order(repeat: Range<i32>) {
-    // Arrange
-    let mut connection = setup_database_connection();
-    reset_test_database(&mut connection);
-
-    // Result
-    for _ in repeat {
-        let output = Command::new("target/debug/fluetl")
-            .args(["import", "order", "--env-file=.env.test"])
-            .output()
-            .expect("Failed to execute command");
-
-        if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            println!("Command executed successfully:\n{}", stdout);
-        } else {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            eprintln!("Command failed:\n{}", stderr);
-        }
-    }
-
-    // Assert
-    let query_results = sql_query("SELECT * FROM `order`")
-        .load::<OrderPlaceholder>(&mut connection)
-        .expect("Failed to query order table");
-
-    let query_result_sample_1 = sql_query("SELECT * FROM `order` WHERE id_order = 1113194")
-        .load::<OrderPlaceholder>(&mut connection)
-        .expect("Failed to query order table");
-
-    let query_result_sample_2 = sql_query("SELECT * FROM `order` WHERE id_order = 1118643")
-        .load::<OrderPlaceholder>(&mut connection)
-        .expect("Failed to query order table");
-
-    assert_eq!(query_results.len(), 5); //No validation error, all 5 rows are inserted
-    assert_eq!(
-        query_result_sample_1[0],
-        OrderPlaceholder {
-            id_order: 1113194,
-            id_client: 1009681,
-            client_name: Some("Client 1".to_string()),
-            order_ref: "OV426946".to_string(),
-            date: chrono::NaiveDateTime::new(
-                NaiveDate::from_ymd_opt(2023, 6, 7).unwrap(),
-                NaiveTime::MIN
-            ),
-            origin: Some("Web".to_string()),
-            order_status: Some("IP".to_string()),
-            completion: Some(1), // 0.9% rounded up to 1
-            po_ref: Some("P23HA01525".to_string())
-        }
-    );
-    assert_eq!(
-        query_result_sample_2[0],
-        OrderPlaceholder {
-            id_order: 1118643,
-            id_client: 1010265,
-            client_name: Some("Client 5".to_string()),
-            order_ref: "OV427619".to_string(),
-            date: chrono::NaiveDateTime::new(
-                NaiveDate::from_ymd_opt(2023, 3, 17).unwrap(),
-                NaiveTime::MIN
-            ),
-            origin: None,
-            order_status: None,
-            completion: Some(100),
-            po_ref: Some("WEB73714".to_string())
-        }
-    );
-}
-
-#[derive(QueryableByName, Debug, PartialEq)]
-struct OrderPlaceholder {
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_order: u32,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_client: u32,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub client_name: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Varchar)]
-    pub order_ref: String,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::Unsigned<diesel::sql_types::Integer>>)]
-    pub completion: Option<u32>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub po_ref: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub origin: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Timestamp)]
-    pub date: chrono::NaiveDateTime,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub order_status: Option<String>,
-}
-
-// ****************** //
-// test fluetl import orderline --env-file=.env.test
-// ****************** //
-#[test]
-fn import_order_line_once() {
-    import_order_line(0..1);
-}
-
-#[test]
-fn import_order_line_10_times() {
-    import_order_line(0..10);
-}
-
-fn import_order_line(repeat: Range<i32>) {
-    // Arrange
-    let mut connection = setup_database_connection();
-    reset_test_database(&mut connection);
-
-    insert_raw_sql("tests/fixtures/languages.sql", &mut connection)
-        .expect("Failed to insert languages.sql fixture file");
-
-    // Import Orders:
-    Command::new("target/debug/fluetl")
-        .args(["import", "order", "--env-file=.env.test"])
-        .output()
-        .expect("Failed to execute command");
-
-    // Result
-    for _ in repeat {
-        let output = Command::new("target/debug/fluetl")
-            .args(["import", "orderline", "--env-file=.env.test"])
-            .output()
-            .expect("Failed to execute command");
-
-        if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            println!("Command executed successfully:\n{}", stdout);
-        } else {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            eprintln!("Command failed:\n{}", stderr);
-        }
-    }
-
-    assert_order_line_results(&mut connection);
-}
-// ****************** //
-// test fluetl import orderline -b -s2 --env-file=.env.test
-// ****************** //
-#[test]
-fn import_order_line_batch() {
-    // Arrange
-    let mut connection = setup_database_connection();
-    reset_test_database(&mut connection);
-
-    insert_raw_sql("tests/fixtures/languages.sql", &mut connection)
-        .expect("Failed to insert languages.sql fixture file");
-
-    // Import Orders:
-    Command::new("target/debug/fluetl")
-        .args(["import", "order", "--env-file=.env.test"])
-        .output()
-        .expect("Failed to execute command");
-
-    let output = Command::new("target/debug/fluetl")
-        .args(["import", "orderline", "-b", "-s2", "--env-file=.env.test"])
-        .output()
-        .expect("Failed to execute command");
-
-    if output.status.success() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        println!("Command executed successfully:\n{}", stdout);
-    } else {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        eprintln!("Command failed:\n{}", stderr);
-    }
-    assert_order_line_results(&mut connection);
-}
-
-fn assert_order_line_results(connection: &mut DbConnection) {
-    // Assert Order Line Table
-    let order_line_query_results = sql_query("SELECT * FROM `order_line`")
-        .load::<OrderLinePlaceholder>(connection)
-        .expect("Failed to query order_line table");
-
-    let order_line_query_result_sample_1 =
-        sql_query("SELECT * FROM `order_line` WHERE id_order_line = 1833200")
-            .load::<OrderLinePlaceholder>(connection)
-            .expect("Failed to query order_line table");
-
-    let order_line_query_result_sample_2 =
-        sql_query("SELECT * FROM `order_line` WHERE id_order_line = 1833202")
-            .load::<OrderLinePlaceholder>(connection)
-            .expect("Failed to query order_line table");
-
-    assert_eq!(order_line_query_results.len(), 3); // 2 must fail because of validation error
-    assert_eq!(
-        order_line_query_result_sample_1[0],
-        OrderLinePlaceholder {
-            id_order_line: 1833200,
-            id_order: 1113194,
-            product_ref: "9995269".to_string(),
-            qty_ordered: 10,
-            qty_reserved: 5,
-            qty_delivered: 5,
-            due_date: Some(NaiveDate::from_ymd_opt(2023, 6, 7).unwrap())
-        }
-    );
-    assert_eq!(
-        order_line_query_result_sample_2[0],
-        OrderLinePlaceholder {
-            id_order_line: 1833202,
-            id_order: 1112737,
-            product_ref: "CDE20".to_string(),
-            qty_ordered: 30,
-            qty_reserved: 15,
-            qty_delivered: 15,
-            due_date: None
-        }
-    );
-
-    // Assert Order Line Lang Table
-    let order_line_item_query_results = sql_query("SELECT * FROM `order_line_lang`")
-        .load::<OrderLineItemPlaceholder>(connection)
-        .expect("Failed to query order_line_lang table");
-
-    let order_line_item_query_result_sample_1 =
-        sql_query("SELECT * FROM `order_line_lang` WHERE id_order_line = 1833200")
-            .load::<OrderLineItemPlaceholder>(connection)
-            .expect("Failed to query order_line table");
-
-    let order_line_item_query_result_sample_2 =
-        sql_query("SELECT * FROM `order_line_lang` WHERE id_order_line = 1833202")
-            .load::<OrderLineItemPlaceholder>(connection)
-            .expect("Failed to query order_line table");
-
-    assert_eq!(order_line_item_query_results.len(), 4); // 1 with es_CO locale and 1 link to an imcomplete order must fail
-    assert_eq!(
-        order_line_item_query_result_sample_1,
-        vec![
-            OrderLineItemPlaceholder {
-                id_order_line: 1833200,
-                id_lang: 1,
-                product_name:
-                    "EN Panier de skimmer GM Standard Liner/Béton + anse - Nouveau Design (ASTRAL)"
-                        .to_string()
-            },
-            OrderLineItemPlaceholder {
-                id_order_line: 1833200,
-                id_lang: 2,
-                product_name:
-                    "FR Panier de skimmer GM Standard Liner/Béton + anse - Nouveau Design (ASTRAL)"
-                        .to_string()
-            }
-        ]
-    );
-    assert_eq!(
-        order_line_item_query_result_sample_2,
-        vec![
-            OrderLineItemPlaceholder {
-                id_order_line: 1833202,
-                id_lang: 1,
-                product_name:
-                    "En OBSOLETE - Volet SK Pisc. Hors-Sol - 105D063 - Remplacé par 4411030601 (SNTE)"
-                        .to_string()
-            },
-            OrderLineItemPlaceholder {
-                id_order_line: 1833202,
-                id_lang: 2,
-                product_name:
-                    "FR OBSOLETE - Volet SK Pisc. Hors-Sol - 105D063 - Remplacé par 4411030601 (SNTE)"
-                        .to_string()
-            }
-        ]
-    );
-}
-
-#[derive(QueryableByName, Debug, PartialEq)]
-struct OrderLinePlaceholder {
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_order_line: u32,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_order: u32,
-    #[diesel(sql_type = diesel::sql_types::Varchar)]
-    pub product_ref: String,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub qty_ordered: u32,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub qty_reserved: u32,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub qty_delivered: u32,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::Date>)]
-    pub due_date: Option<chrono::NaiveDate>,
-}
-
-#[derive(QueryableByName, Debug, PartialEq)]
-struct OrderLineItemPlaceholder {
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_order_line: u32,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_lang: u32,
-    #[diesel(sql_type = diesel::sql_types::Varchar)]
-    pub product_name: String,
-}
-
-// ****************** //
-// test fluetl import delivery-slip --env-file=.env.test
-// ****************** //
-#[test]
-fn import_delivery_slip_once() {
-    import_delivery_slip(0..1);
-}
-
-#[test]
-fn import_delivery_slip_10_times() {
-    import_delivery_slip(0..10);
-}
-
-fn import_delivery_slip(repeat: Range<i32>) {
-    // Arrange
-    let mut connection = setup_database_connection();
-    reset_test_database(&mut connection);
-
-    // Result
-    for _ in repeat {
-        let output = Command::new("target/debug/fluetl")
-            .args(["import", "delivery-slip", "--env-file=.env.test"])
-            .output()
-            .expect("Failed to execute command");
-
-        if output.status.success() {
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            println!("Command executed successfully:\n{}", stdout);
-        } else {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            eprintln!("Command failed:\n{}", stderr);
-        }
-    }
-
-    assert_delivery_slip_table(&mut connection);
-}
-// ****************** //
-// test fluetl import delivery-slip -b -s2 --env-file=.env.test
-// ****************** //
-#[test]
-fn import_delivery_slip_batch() {
-    // Arrange
-    let mut connection = setup_database_connection();
-    reset_test_database(&mut connection);
-
-    let output = Command::new("target/debug/fluetl")
-        .args([
-            "import",
-            "delivery-slip",
-            "-b",
-            "-s2",
-            "--env-file=.env.test",
-        ])
-        .output()
-        .expect("Failed to execute command");
-
-    if output.status.success() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        println!("Command executed successfully:\n{}", stdout);
-    } else {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        eprintln!("Command failed:\n{}", stderr);
-    }
-
-    assert_delivery_slip_table(&mut connection);
-}
-
-fn assert_delivery_slip_table(connection: &mut DbConnection) {
-    // Assert
-    let query_results = sql_query("SELECT * FROM `delivery_slip`")
-        .load::<DeliverySlipPlaceholder>(connection)
-        .expect("Failed to query delivery_slip table");
-
-    let query_result_sample_1 =
-        sql_query("SELECT * FROM `delivery_slip` WHERE id_delivery_slip = 1011124")
-            .load::<DeliverySlipPlaceholder>(connection)
-            .expect("Failed to query delivery_slip table");
-
-    let query_result_sample_2 =
-        sql_query("SELECT * FROM `delivery_slip` WHERE id_delivery_slip = 1012515")
-            .load::<DeliverySlipPlaceholder>(connection)
-            .expect("Failed to query delivery_slip table");
-
-    let query_result_sample_3 =
-        sql_query("SELECT * FROM `delivery_slip` WHERE id_delivery_slip = 1008692")
-            .load::<DeliverySlipPlaceholder>(connection)
-            .expect("Failed to query delivery_slip table");
-
-    assert_eq!(query_results.len(), 4); // 1 must fail because of csv error
-    assert_eq!(
-        query_result_sample_1[0],
-        DeliverySlipPlaceholder {
-            id_delivery_slip: 1011124,
-            id_client: 1009681,
-            reference: "BL501714".to_string(),
-            shipping_date: Some(NaiveDate::from_ymd_opt(2023, 3, 1).unwrap()),
-            po_ref: Some("ABC / DEF".to_string()),
-            carrier_name: Some("UPS".to_string()),
-            status: Some("1".to_string()),
-            tracking_number: Some("250238315641646373".to_string()),
-            tracking_link: Some("https://tracking1.com/123".to_string())
-        }
-    );
-    assert_eq!(
-        query_result_sample_2[0],
-        DeliverySlipPlaceholder {
-            id_delivery_slip: 1012515,
-            id_client: 1012489,
-            reference: "BL501392".to_string(),
-            shipping_date: None,
-            po_ref: None,
-            carrier_name: None,
-            status: None,
-            tracking_number: None,
-            tracking_link: None
-        }
-    );
-    assert_eq!(
-        query_result_sample_3[0],
-        DeliverySlipPlaceholder {
-            id_delivery_slip: 1008692,
-            id_client: 1009681,
-            reference: "BL501550".to_string(),
-            shipping_date: Some(NaiveDate::from_ymd_opt(2023, 3, 1).unwrap()),
-            po_ref: Some("ACALOG201138802".to_string()),
-            carrier_name: Some("Carrier 1".to_string()),
-            status: Some("1".to_string()),
-            tracking_number: Some("TrackingNo3".to_string()),
-            tracking_link: None
-        }
-    );
-}
-
-#[derive(QueryableByName, Debug, PartialEq)]
-struct DeliverySlipPlaceholder {
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_delivery_slip: u32,
-    #[diesel(sql_type = diesel::sql_types::Unsigned<diesel::sql_types::Integer>)]
-    pub id_client: u32,
-    #[diesel(sql_type = diesel::sql_types::Varchar)]
-    pub reference: String,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::Date>)]
-    pub shipping_date: Option<chrono::NaiveDate>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub po_ref: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub carrier_name: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub status: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub tracking_number: Option<String>,
-    #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::VarChar>)]
-    pub tracking_link: Option<String>,
-}
+mod cli_sub_routines;
 
 // ****************** //
 // Helper functions
