+++ tests/cli_sub_routines/mod.rs
@@ -1,0 +1,4 @@
+mod delivery_slip;
+mod mapping_client;
+mod order;
+mod order_line;
